+++ src/kmeans_stream.h
@@ -10,12 +10,11 @@ typedef struct{
     int ncols;      // number of features
     int k;          // number of clusters
     int n_seen;     // total points seen so far (needed by warmup)
+    double halflife;// -1.0 = count-based, >0 = fixed learnign rate 
 
 } KMeansStreamModel;
 
-void find_closest_centroid();
-void update_centroid();
-SEXP C_kmeans_stream_new(SEXP k_r, SEXP d_r);
+SEXP C_kmeans_stream_new(SEXP k_r, SEXP d_r, SEXP halflife);
 SEXP C_kmeans_stream_update(SEXP ptr_r, SEXP x_r);
 SEXP C_kmeans_stream_predict(SEXP ptr_r, SEXP x_r);