+++ src/utils.c
@@ -6,7 +6,7 @@
 #include "utils.h"
 
 /* Computes euclidean distance between two vectors of the same size */
-double euclidean_distance(double* a, double* b, size_t n){
+double euclidean_distance(const double* a, const double* b, size_t n){
     double sum = 0.0;
     for(size_t i=0; i<n; i++){
         sum += pow(a[i]-b[i],2);