+++ src/utils.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-double euclidean_distance(double* a, double* b, size_t n);
+double euclidean_distance(const double* a, const double*, size_t n);
 void col_to_row_major(const double *src, double *dst, int nrows, int ncols);
 void row_to_col_major(const double *src, double *dst, int nrows, int ncols);